added validation to favorites

+++ src/main/java/com/example/demo/services/FavoriteService.java
@@ -1,5 +1,6 @@
 package com.example.demo.services;
 
+import com.example.demo.dto.response.FavoriteDTO;
 import com.example.demo.exceptions.duplicate.ProductDuplicateException;
 import com.example.demo.exceptions.notfound.ProductNotFoundException;
 import com.example.demo.models.Favorite;
@@ -7,13 +8,13 @@ import com.example.demo.models.Product;
 import com.example.demo.repositories.FavoriteRepository;
 import com.example.demo.repositories.MyUserRepository;
 import com.example.demo.repositories.ProductRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,15 +24,27 @@ public class FavoriteService {
     private final ProductRepository productRepository;
     private final MyUserRepository userRepository;
 
-    public List<Product> getUserFavorites(String email) {
-        return favoriteRepository.findByUserEmail(email).stream()
-                .map(Favorite::getProduct)
+    public List<FavoriteDTO> getUserFavorites(String email) {
+        return favoriteRepository.findByUserEmail(email)
+                .stream()
+                .map(this::convertToDto)
                 .collect(Collectors.toList());
     }
 
+    @Transactional
     public String addFavorite(String email, Long productId) {
-        Product product = existingProduct(productId);
-        isProductDuplicated(email, productId);
+        Product product = productRepository.findById(productId)
+                .orElseThrow(() -> {
+                    Map<String, List<String>> error = new HashMap<>();
+                    error.put("product", List.of("Product with id " + productId + " not found"));
+                    return new ProductNotFoundException(error);
+                });
+
+        if (favoriteRepository.existsByUserEmailAndProductId(email, productId)) {
+            Map<String, List<String>> errors = new HashMap<>();
+            errors.put("favorite", List.of("Product with id: " + productId + " already in favorites"));
+            throw new ProductDuplicateException(errors);
+        }
 
         Favorite favorite = Favorite.builder()
                 .user(userRepository.findByEmail(email).orElseThrow())
@@ -43,25 +56,24 @@ public class FavoriteService {
         return "Added to favorites";
     }
 
-    private void isProductDuplicated(String email, Long productId) {
-        if (favoriteRepository.existsByUserEmailAndProductId(email, productId)) {
+    public void removeFavorite(String email, Long productId) {
+        if (!productRepository.existsById(productId)) {
+            throw new ProductNotFoundException(Map.of("product", List.of("Product not found")));
+        }
+
+        if (!favoriteRepository.existsByUserEmailAndProductId(email, productId)) {
             Map<String, List<String>> errors = new HashMap<>();
-            errors.put("favorite", List.of("Product with id: " + productId + " already in favorites"));
+            errors.put("favorite", List.of("Product with id: " + productId + " not yet in favorites"));
             throw new ProductDuplicateException(errors);
         }
-    }
-
-    private Product existingProduct(Long productId) {
-        return productRepository.findById(productId)
-                .orElseThrow(() -> {
-                    Map<String, List<String>> error = new HashMap<>();
-                    error.put("product", List.of("Product with id " + productId + " not found"));
-                    return new ProductNotFoundException(error);
-                });
-    }
-
-    public void removeFavorite(String email, Long productId) {
-        existingProduct(productId);
         favoriteRepository.deleteByUserEmailAndProductId(email, productId);
+    }
+
+    private FavoriteDTO convertToDto(Favorite favorite) {
+        return new FavoriteDTO(
+                favorite.getId(),
+                favorite.getUser().getId(),
+                favorite.getProduct().getId()
+        );
     }
 }
